up
MJEテンプレ作成時もアプデ内容を反映
赤合格の不可の割合をMJE式条件にする機能の修正
TJCの内容を表示する際の内容を更新

+++ JiroPackEditor/FileDrop.cs
@@ -54,7 +54,7 @@ namespace JiroPackEditor {
 			};
 
 			Marshal.StructureToPtr(dropFiles, hGlobal, true);
-			var fuga = hGlobal.ToInt32();
+			var fuga = hGlobal.ToInt64();
 			var hoge = fuga + dropFiles.pFiles;
 			Marshal.Copy(names, 0, new IntPtr(hoge), names.Length);
